Fix pl plotting (#55)
* Fix labels for plots

The previous label keyword had no effect; this fixes that and adds doc string entries for labels.

* Add max_depth attribute to PersLandscapeApprox

* Fix depth axis

The tick marks on the y-axis (corresponding to depth of the landscapes) were completely off and mislabeled.

* Add 'depth_range' parameter to signature

The new 'depth_range' parameter allows the user to specify a range of depths to be plotted.
Omitting it plots all depths.

* Clean up doc strings and signatures

The doc strings were previously in the helper functions, which were not
publicly accessible. Move the docstrings into the two main plotting functions
and clean up the signatures of the plotting functions.

* Fix padding

The previous way of padding the figures changed their values slightly. The 'padding' parameter now
adds a slight margin to the plot, especially useful for 2-d 'simple' plots.

* Remove depth_spacing/depth_padding

`depth_spacing` caused overcrowding issues with labelling. If it needs to be modified, it should be
done on a case by case basis by passing an axis.

* Fix "gca('3d') deprecation warning"

* Update notebooks